Add category controller and update routes/services
Introduced CategoryController with create, delete, and getAll methods. Updated admin and anime routes to use CategoryController. Refactored repository and service layers for category and genre to format DTOs explicitly and validate ObjectId on delete. Adjusted server to use new route structure.

+++ src/service/category.service.ts
@@ -1,5 +1,6 @@
 import { CategoryDTO } from "@/interface/dto/category.dto";
 import CategoryRepository from "@/repository/category.repository";
+import { isObjectIdOrHexString } from "mongoose";
 
 class CategoryService {
   constructor(private categoryRepository = new CategoryRepository()) {}
@@ -28,6 +29,10 @@ class CategoryService {
 
   async deleteCategory(id: string) {
     try {
+      if (!isObjectIdOrHexString(id)) {
+        return null;
+      }
+
       return await this.categoryRepository.delete(id);
     } catch (error) {
       if (error instanceof Error) {
@@ -57,4 +62,4 @@ class CategoryService {
   }
 }
 
-export default new CategoryService();
+export default CategoryService;
